requirements.txt
# PPTX-to-Podcast Converter
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PPTX Parsing
python-pptx>=0.6.21
Pillow>=10.0.0

# AI / LLM
ollama>=0.1.0

# Text-to-Speech
TTS>=0.22.0
torch>=2.0.0

# Audio Processing
mutagen>=1.47.0

# Utilities
python-multipart>=0.0.6
